Rename _club.h to _memedit.h
